fixing delete button on consultant

+++ addcomment.php
@@ -99,7 +99,13 @@ if(isset($_SESSION['username']) && $dta['sess']==$_SESSION['username'])
 			$insq= $conn->prepare($qc);
 			$insq->bindValue( ":comment", $comment, PDO::PARAM_STR );
 			$insq->execute();
+			if($com_type ==7) {
 			echo "<script>alert('Comment Added.');window.location.href='consultantcmd.php?do=delete&id=$com_postid'</script>";
+			}
+			else {
+				echo "<script>alert('Comment Added.');window.close();</script>";
+			
+			}
 		}
 		else{
 			?>
